audit-service/requirements.txt
Flask==2.2.3
boto3==1.26.20
pymysql==1.0.2
requests==2.31.0
python-dotenv==1.0.0
pika==1.3.1
